feat: Data retrieval and persistence
I created a re-usable/generic repositories for retrieval and persistence of entities.

This, for me, is like a CQRS pattern, in a much simpler way but still CQRS is the real deal for an enterprise level application.

Also, added my custom database exception for much easier tracing of expected errors.